Remove __init__.py and rename package to avoid conflicts with top-level one
