+++ runtimes/mlflow/tests/conftest.py
@@ -12,7 +12,7 @@ from mlserver.types import InferenceRequest
 
 from mlserver_mlflow import MLflowRuntime
 
-from .fixtures import MNISTDataModule, LightningMNISTClassifier
+from torch_fixtures import MNISTDataModule, LightningMNISTClassifier
 
 TESTS_PATH = os.path.dirname(__file__)
 TESTDATA_PATH = os.path.join(TESTS_PATH, "testdata")
